requirements.txt
Flask
Pillow
numpy
webcolors
